requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.8
inexactsearch==1.0.2
pyspellchecker==0.8.1
requests==2.32.3
setuptools==74.0.0
silpa_common==0.3
soundex==1.1.3
soupsieve==2.6
urllib3==2.2.2
